requirements.txt
python-dotenv==1.0.1
together==1.2.1
flask==3.0.3